Just use sema4ai-data.dataserver.status to get data server status and stop using sema4ai-data.dataserver.start command since its deprecated.
