+++ sema4ai/vscode-client/src/rcc.ts
@@ -148,7 +148,7 @@ async function downloadRcc(
             throw new Error("Currently only Linux amd64 is supported.");
         }
     }
-    const RCC_VERSION = "v18.5.0";
+    const RCC_VERSION = "v20.1.1";
     const prefix = "https://cdn.sema4.ai/rcc/releases/" + RCC_VERSION;
     const url: string = prefix + relativePath;
     return await downloadWithProgress(url, progress, token, location);
